Moved types to types.go file and other changes.
