+++ servers/provisioner/utils/types.go
@@ -3,6 +3,10 @@
 // Package utils
 package utils
 
+import (
+	"encoding/xml"
+)
+
 const (
 	ACCESS_POLICY_PREFIX = "acp_"
 	USER_PREFIX          = "user_"
@@ -26,4 +30,51 @@ type IAMCredentials struct {
 	SystemId          string
 	Endpoint          string
 	Proxy             string
+}
+
+
+// Defines BucketRequest structure with options (Versioning, Locking, Compression)
+type BucketRequest struct {
+	Compression     bool   `json:"Compression"`
+	Versioning      string `json:"Versioning,omitempty"`
+	Locking         string `json:"Locking,omitempty"`
+	RetentionMode   string `json:"RetentionMode,omitempty"`
+	ObjectLockDays  int    `json:"ObjectLockDays,omitempty"`
+	ObjectLockYears int    `json:"ObjectLockYears,omitempty"`
+}
+
+// Add new structs for S3 bucket creation
+type SpaceQuota struct {
+	QuotaType     string `json:"QuotaType"`
+	QuotaLimitMiB int    `json:"QuotaLimitMiB"`
+}
+
+type CreateBucketRequest struct {
+	LocationConstraint string     `json:"LocationConstraint"`
+	Compression        bool       `json:"Compression"`
+	BucketPolicy       string     `json:"BucketPolicy"`
+	Versioning         string     `json:"Versioning,omitempty"`
+	ObjectLockEnabled  string     `json:"ObjectLockEnabled,omitempty"`
+	ObjectLockMode     string     `json:"ObjectLockMode,omitempty"`
+	ObjectLockDays     int        `json:"ObjectLockDays,omitempty"`
+	ObjectLockYears    int        `json:"ObjectLockYears,omitempty"`
+	SpaceQuota         SpaceQuota `json:"SpaceQuota"`
+}
+
+// ObjectLockConfiguration represents the XML structure for object lock configuration
+type ObjectLockConfiguration struct {
+	XMLName           xml.Name        `xml:"ObjectLockConfiguration"`
+	Xmlns             string          `xml:"xmlns,attr"`
+	ObjectLockEnabled string          `xml:"ObjectLockEnabled"`
+	Rule              *ObjectLockRule `xml:"Rule,omitempty"`
+}
+
+type ObjectLockRule struct {
+	DefaultRetention ObjectLockDefaultRetention `xml:"DefaultRetention"`
+}
+
+type ObjectLockDefaultRetention struct {
+	Mode  string `xml:"Mode,omitempty"`
+	Days  int    `xml:"Days,omitempty"`
+	Years int    `xml:"Years,omitempty"`
 }
